+++ idf_build_apps/__init__.py
@@ -7,7 +7,7 @@ Tools for building ESP-IDF related apps.
 
 import logging
 
-__version__ = '1.0.2'
+__version__ = '1.0.3'
 
 LOGGER = logging.getLogger('idf_build_apps')